Move test helpers into their own files
