+++ LumiformCaseStudy/LumiformTests/Form API/URLSessionHTTPClientTests.swift
@@ -127,77 +127,4 @@ final class URLSessionHTTPClientTests: XCTestCase {
     private func anyHTTPURLResponse() -> HTTPURLResponse {
         HTTPURLResponse(url: anyURL(), statusCode: 200, httpVersion: nil, headerFields: nil)!
     }
-
-    private class URLProtocolStub: URLProtocol {
-
-        private struct Stub {
-            let data: Data?
-            let response: URLResponse?
-            let error: Error?
-            let requestObserver: ((URLRequest) -> Void)?
-        }
-
-        private static let queue = DispatchQueue(label: "URLProtocolStub.queue")
-
-        private static var _stub: Stub?
-        private static var stub: Stub? {
-            get { queue.sync { _stub } }
-            set { queue.sync { _stub = newValue } }
-        }
-
-        class func observeRequests(_ completion: @escaping (URLRequest) -> Void) {
-            stub = Stub(data: nil, response: nil, error: nil, requestObserver: completion)
-        }
-
-        class func stub(data: Data?, response: URLResponse?, error: Error?) {
-            stub = Stub(data: data, response: response, error: error, requestObserver: nil)
-        }
-
-        class func removeStub() {
-            stub = nil
-        }
-
-        // MARK: - URLProtocol
-
-        override class func canInit(with request: URLRequest) -> Bool {
-            true
-        }
-        
-        override class func canonicalRequest(for request: URLRequest) -> URLRequest {
-            request
-        }
-
-        override func startLoading() {
-            guard let stub = URLProtocolStub.stub else { return }
-
-            if let data = stub.data {
-                client?.urlProtocol(self, didLoad: data)
-            }
-
-            if let response = stub.response {
-                client?.urlProtocol(self, didReceive: response, cacheStoragePolicy: .notAllowed)
-            }
-
-            if let error = stub.error {
-                client?.urlProtocol(self, didFailWithError: error)
-            }
-
-            client?.urlProtocolDidFinishLoading(self)
-            stub.requestObserver?(request)
-        }
-
-        override func stopLoading() {}
-    }
-}
-
-func anyURL() -> URL {
-    URL(string: "https://some-url.com")!
-}
-
-func anyNSError() -> NSError {
-    NSError(domain: "someError", code: 0)
-}
-
-func anyData() -> Data {
-    Data("some data".utf8)
 }
